legacy/requirements.txt
# Core dependencies
gitpython>=3.1.0
click>=8.0.0
rich>=13.0.0
psycopg2-binary>=2.9.0
flask>=2.3.0
flask-cors>=4.0.0
requests>=2.31.0
schedule>=1.2.0

# Data processing (optional - can be excluded for Python 3.13+)
pandas>=2.2.0
numpy>=1.26.0

# Development and testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0

# Documentation
sphinx>=7.0.0
sphinx-rtd-theme>=1.3.0
python-dotenv>=1.0.0
